Bringing features back in line so the features module is happy.
We were defining the views.inc and default_views files in a separate
"views" folder, but this required removing manually the hook_views_api
call within the features file, which was causing the Default/Overrideen
check to poop out and causing general confusion. We're still keeping all
the handlers in the "views" folders tho.